requirements.txt
Flask
pillow
opencv-python
werkzeug
numpy==1.19.3